pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-clabe"
description = "A library for a minimal framework that can be used to build experimental interfaces."
authors = [
    { name = "Bruno Cruz", email = "bruno.cruz@alleninstitute.org" }
]
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dynamic = ["version", "readme"]

classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: Microsoft :: Windows"
]

dependencies = [
    "pydantic>=2.7",
    "pydantic-settings",
    "gitpython",
    "semver",
    "rich",
    "aind_behavior_services < 1",
]

[project.urls]
Documentation = "https://allenneuraldynamics.github.io/clabe/"
Repository = "https://github.com/AllenNeuralDynamics/clabe/"
Issues = "https://github.com/AllenNeuralDynamics/clabe/issues"
Changelog = "https://github.com/AllenNeuralDynamics/clabe/releases"

[project.optional-dependencies]
aind-services = [
    "aind-slims-api",
    "aind-watchdog-service",
    "aind-data-schema<2",
    "ms-active-directory",
    "cryptography",
    "winkerberos; sys_platform == 'win32'",
    "ldap3; sys_platform == 'win32'",
]

dev = [
    "ruff",
    "codespell",
    "coverage",
    "aind-clabe[aind-services]",
]

docs = [
    "aind-clabe[aind-services]",
    'mkdocs',
    'mkdocs-material',
    'mkdocstrings[python]',
    'pymdown-extensions',
    'ruff',
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aind_behavior_task_launcher = ["py.typed"]

[tool.setuptools.dynamic]
version = { attr = "clabe.__version__" }
readme = { file = "README.md", content-type = "text/markdown" }

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["tests/assets/*"]

[tool.ruff.lint]
extend-select = ["Q", "RUF100", "C90", "I"]
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = "google" }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock'
ignore-words-list = "nd, assertIn"

[tool.coverage.run]
omit = ["*__init__*"]
source = ["clabe", "tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "from",
    "import",
    "pragma: no cover"
]
fail_under = 75